gui: Merge in gui_events
