+++ gui.h
@@ -50,5 +50,12 @@ void     saveCurrentPic();
 
 void     evt_chooseDir     ();
 void     evt_show_about    ();
+void     evt_AbortRx       ();
+void     evt_changeDevices ();
+void     evt_clearPix      ();
+void     evt_clickimg      ();
+void     evt_deletewindow  ();
+void     evt_GetAdaptive   ();
+void     evt_ManualStart   ();
 
 #endif
